+++ common/src/main/java/dev/terminalmc/chatnotify/config/Config.java
@@ -18,6 +18,7 @@ package dev.terminalmc.chatnotify.config;
 
 import com.google.gson.*;
 import dev.terminalmc.chatnotify.ChatNotify;
+import dev.terminalmc.chatnotify.platform.Services;
 import dev.terminalmc.chatnotify.util.JsonUtil;
 import net.minecraft.sounds.SoundSource;
 import org.jetbrains.annotations.NotNull;
@@ -50,7 +51,7 @@ import java.util.function.Supplier;
 public class Config {
     public static final int VERSION = 9;
     public final int version = VERSION;
-    private static final Path DIR_PATH = Path.of("config");
+    private static final Path CONFIG_DIR = Services.PLATFORM.getConfigDir();
     public static final String FILE_NAME = ChatNotify.MOD_ID + ".json";
     public static final String UNREADABLE_FILE_NAME = ChatNotify.MOD_ID + ".unreadable.json";
     public static final String OLD_FILE_NAME = ChatNotify.MOD_ID + ".old.json";
@@ -316,7 +317,7 @@ public class Config {
     // Load and save
 
     public static @NotNull Config load() {
-        Path file = DIR_PATH.resolve(FILE_NAME);
+        Path file = CONFIG_DIR.resolve(FILE_NAME);
         Config config = null;
         if (Files.exists(file)) {
             JsonUtil.reset();
@@ -347,8 +348,8 @@ public class Config {
     private static void backup(String path) {
         try {
             ChatNotify.LOG.warn("Copying {} to {}", FILE_NAME, path);
-            if (!Files.isDirectory(DIR_PATH)) Files.createDirectories(DIR_PATH);
-            Path file = DIR_PATH.resolve(FILE_NAME);
+            if (!Files.isDirectory(CONFIG_DIR)) Files.createDirectories(CONFIG_DIR);
+            Path file = CONFIG_DIR.resolve(FILE_NAME);
             Path backupFile = file.resolveSibling(path);
             Files.move(file, backupFile, StandardCopyOption.ATOMIC_MOVE,
                     StandardCopyOption.REPLACE_EXISTING);
@@ -361,8 +362,8 @@ public class Config {
         if (instance == null) return;
         instance.validate();
         try {
-            if (!Files.isDirectory(DIR_PATH)) Files.createDirectories(DIR_PATH);
-            Path file = DIR_PATH.resolve(FILE_NAME);
+            if (!Files.isDirectory(CONFIG_DIR)) Files.createDirectories(CONFIG_DIR);
+            Path file = CONFIG_DIR.resolve(FILE_NAME);
             Path tempFile = file.resolveSibling(file.getFileName() + ".tmp");
             try (OutputStreamWriter writer = new OutputStreamWriter(
                     new FileOutputStream(tempFile.toFile()), StandardCharsets.UTF_8)) {